+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * added browser and OS to filter options
 * added weekday mode configuration to filter
 * added the option to add multiple IP address filters
+* added simple flat IP blocklist implementation
 * removed the requirement for event to be set to filter for key/value options
 * updated dependencies